# Remove early indices