--- a/code/factorbase/src/main/java/ca/sfu/cs/factorbase/tables/Sort_merge3.java
+++ b/code/factorbase/src/main/java/ca/sfu/cs/factorbase/tables/Sort_merge3.java
@@ -188,7 +188,7 @@ public class Sort_merge3 {
             long time4 = System.currentTimeMillis();
 //            System.out.print("\t insert time: " + (time4 - time3));
             st2.execute("DROP TABLE IF EXISTS " + table3 + ";");
-            st2.execute("CREATE TABLE " + table3 + " LIKE " + table1 + ";");
+            st2.execute("CREATE TABLE " + table3 + " SELECT * FROM " + table1 + " LIMIT 0;");
             st2.execute("LOAD DATA LOCAL INFILE 'sort_merge.csv' INTO TABLE " + table3 + " FIELDS TERMINATED BY '$' LINES TERMINATED BY '\\n';");
 
             rst1.close();
@@ -207,7 +207,7 @@ public class Sort_merge3 {
         } else { // Aug 18, 2014 zqian: Handle the extreme case when there's only `mult` column.
             logger.fine("\n \t Handle the extreme case when there's only `mult` column \n");
             st2.execute("DROP TABLE IF EXISTS " + table3 + ";");
-            st2.execute("CREATE TABLE " + table3 + " LIKE " + table1 + ";");
+            st2.execute("CREATE TABLE " + table3 + " SELECT * FROM " + table1 + " LIMIT 0;");
             logger.fine("INSERT INTO " + table3 + " SELECT (" + table1 + ".mult - " + table2 + ".mult) AS mult FROM " + table1 + ", " + table2 + ";");
             st2.execute("INSERT INTO " + table3 + " SELECT (" + table1 + ".mult - " + table2 + ".mult) AS mult FROM " + table1 + ", " + table2 + ";");
         }