We were creating indices before inserting data into the output table for sort merge, which was slowing the performance down a little bit, so this logic has been removed.  This change also fixes the duplicated indices we were seeing in the false tables.  Here is an example of duplicated indices:
```
BEFORE CHANGES:
mysql> show index from `a_star`;
+--------+------------+----------+--------------+------------------------+-----------+-------------+----------+--------+------+------------+---------+---------------+
| Table  | Non_unique | Key_name | Seq_in_index | Column_name            | Collation | Cardinality | Sub_part | Packed | Null | Index_type | Comment | Index_comment |
+--------+------------+----------+--------------+------------------------+-----------+-------------+----------+--------+------+------------+---------+---------------+
| a_star |          1 | a_star   |            1 | MULT                   | A         |          21 |     NULL | NULL   |      | BTREE      |         |               |
| a_star |          1 | a_star   |            2 | intelligence(student0) | A         |          21 |     NULL | NULL   | YES  | BTREE      |         |               |
| a_star |          1 | a_star   |            3 | popularity(prof0)      | A         |          21 |     NULL | NULL   | YES  | BTREE      |         |               |
| a_star |          1 | a_star   |            4 | ranking(student0)      | A         |          21 |     NULL | NULL   | YES  | BTREE      |         |               |
| a_star |          1 | a_star   |            5 | teachingability(prof0) | A         |          21 |     NULL | NULL   | YES  | BTREE      |         |               |
+--------+------------+----------+--------------+------------------------+-----------+-------------+----------+--------+------+------------+---------+---------------+
5 rows in set (0.00 sec)

mysql> show index from `a_false`;
+---------+------------+----------+--------------+------------------------+-----------+-------------+----------+--------+------+------------+---------+---------------+
| Table   | Non_unique | Key_name | Seq_in_index | Column_name            | Collation | Cardinality | Sub_part | Packed | Null | Index_type | Comment | Index_comment |
+---------+------------+----------+--------------+------------------------+-----------+-------------+----------+--------+------+------------+---------+---------------+
| a_false |          1 | a_star   |            1 | MULT                   | A         |          21 |     NULL | NULL   |      | BTREE      |         |               |
| a_false |          1 | a_star   |            2 | intelligence(student0) | A         |          21 |     NULL | NULL   | YES  | BTREE      |         |               |
| a_false |          1 | a_star   |            3 | popularity(prof0)      | A         |          21 |     NULL | NULL   | YES  | BTREE      |         |               |
| a_false |          1 | a_star   |            4 | ranking(student0)      | A         |          21 |     NULL | NULL   | YES  | BTREE      |         |               |
| a_false |          1 | a_star   |            5 | teachingability(prof0) | A         |          21 |     NULL | NULL   | YES  | BTREE      |         |               |
| a_false |          1 | a_false  |            1 | MULT                   | A         |          21 |     NULL | NULL   |      | BTREE      |         |               |
| a_false |          1 | a_false  |            2 | intelligence(student0) | A         |          21 |     NULL | NULL   | YES  | BTREE      |         |               |
| a_false |          1 | a_false  |            3 | popularity(prof0)      | A         |          21 |     NULL | NULL   | YES  | BTREE      |         |               |
| a_false |          1 | a_false  |            4 | ranking(student0)      | A         |          21 |     NULL | NULL   | YES  | BTREE      |         |               |
| a_false |          1 | a_false  |            5 | teachingability(prof0) | A         |          21 |     NULL | NULL   | YES  | BTREE      |         |               |
+---------+------------+----------+--------------+------------------------+-----------+-------------+----------+--------+------+------------+---------+---------------+
10 rows in set (0.00 sec)

AFTER CHANGES:
mysql> show index from `a_star`;
+--------+------------+----------+--------------+------------------------+-----------+-------------+----------+--------+------+------------+---------+---------------+
| Table  | Non_unique | Key_name | Seq_in_index | Column_name            | Collation | Cardinality | Sub_part | Packed | Null | Index_type | Comment | Index_comment |
+--------+------------+----------+--------------+------------------------+-----------+-------------+----------+--------+------+------------+---------+---------------+
| a_star |          1 | a_star   |            1 | MULT                   | A         |          21 |     NULL | NULL   |      | BTREE      |         |               |
| a_star |          1 | a_star   |            2 | intelligence(student0) | A         |          21 |     NULL | NULL   | YES  | BTREE      |         |               |
| a_star |          1 | a_star   |            3 | popularity(prof0)      | A         |          21 |     NULL | NULL   | YES  | BTREE      |         |               |
| a_star |          1 | a_star   |            4 | ranking(student0)      | A         |          21 |     NULL | NULL   | YES  | BTREE      |         |               |
| a_star |          1 | a_star   |            5 | teachingability(prof0) | A         |          21 |     NULL | NULL   | YES  | BTREE      |         |               |
+--------+------------+----------+--------------+------------------------+-----------+-------------+----------+--------+------+------------+---------+---------------+
5 rows in set (0.00 sec)

mysql> show index from `a_false`;
+---------+------------+----------+--------------+------------------------+-----------+-------------+----------+--------+------+------------+---------+---------------+
| Table   | Non_unique | Key_name | Seq_in_index | Column_name            | Collation | Cardinality | Sub_part | Packed | Null | Index_type | Comment | Index_comment |
+---------+------------+----------+--------------+------------------------+-----------+-------------+----------+--------+------+------------+---------+---------------+
| a_false |          1 | a_false  |            1 | MULT                   | A         |          21 |     NULL | NULL   |      | BTREE      |         |               |
| a_false |          1 | a_false  |            2 | intelligence(student0) | A         |          21 |     NULL | NULL   | YES  | BTREE      |         |               |
| a_false |          1 | a_false  |            3 | popularity(prof0)      | A         |          21 |     NULL | NULL   | YES  | BTREE      |         |               |
| a_false |          1 | a_false  |            4 | ranking(student0)      | A         |          21 |     NULL | NULL   | YES  | BTREE      |         |               |
| a_false |          1 | a_false  |            5 | teachingability(prof0) | A         |          21 |     NULL | NULL   | YES  | BTREE      |         |               |
+---------+------------+----------+--------------+------------------------+-----------+-------------+----------+--------+------+------------+---------+---------------+
5 rows in set (0.01 sec)
```