@oschulte 2019-04-08
good catch. Does it help with runtime as well?