@rmar3a 2019-04-08
Yes, I posted the runtimes on Slack.  Before the index change Financial took about 67 minutes and after the index change it took about 45 minutes.  As for IMDB MovieLens the runtime for two trials were:
```
real    241m4.809s
real    247m0.220s
```
which I believe is a little bit faster than before, but not as much of an improvement as Financial.