@rmar3a 2019-04-09
Just some follow-up run times I've collected.

Before this change (having the indices before insertion) gives run times like:
```
real    286m2.421s
```
and after the change (removing the indices before insertion) gives run times like:
```
real    253m11.921s
```
So this is a decrease in the runtime for IMDB MovieLens, but not in the same percentage as Financial.