@rmar3a 2019-04-10
More run time information can be found at pull request https://github.com/sfu-cl-lab/FactorBase/pull/108